requirements.txt
arrow==0.12.1
binaryornot==0.4.4
blessings==1.6.1
boto3==1.6.3
botocore==1.9.3
certifi==2018.1.18
chardet==3.0.4
click==6.7
cookiecutter==1.6.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
decorator==4.2.1
dj-database-url==0.5.0
Django==2.0.2
django-configurations==2.0
django-filter==1.1.0
django-model-utils==3.1.1
django-nose==1.4.5
django-storages==1.6.5
django-unique-upload==0.2.1
djangorestframework==3.7.7
docutils==0.14
factory-boy==2.10.0
Faker==0.8.11
flake8==3.5.0
future==0.16.0
gunicorn==19.7.1
idna==2.6
ipdb==0.11
ipython==6.2.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.11.1
Jinja2==2.10
jinja2-time==0.2.0
jmespath==0.9.3
livereload==2.5.1
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
mkdocs==0.17.2
mock==2.0.0
newrelic==2.106.0.87
nose==1.3.7
nose-progressive==1.5.1
parso==0.1.1
pbr==3.1.1
pexpect==4.4.0
pickleshare==0.7.4
poyo==0.4.1
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2018.3
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.13
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.2
tornado==4.5.3
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
whichcraft==0.4.1
